fix: api/goods/route.ts에서 mappingResponse와 orderingData를 utils 로 이동시킴으로써 빌드 오류 해결
