+++ src/app/api/goods/route.ts
@@ -1,12 +1,6 @@
-import {
-  GetGoodsRequestParams,
-  GetGoodsResponse,
-  Order,
-} from "@/features/main/api/getGoods";
-import { RawGoodDTO } from "@/shared/api/house/types/item";
+import { GetGoodsRequestParams, Order } from "@/features/main/api/getGoods";
 import { goods } from "@/shared/mock-data/goods";
-import { numberRounded } from "@/shared/utils/data";
-import { formatNumberWithCommas } from "@/shared/utils/format/number";
+import { mappingGoodsResponse, orderingGoodsData } from "@/shared/utils/data";
 import { NextRequest } from "next/server";
 
 export const GET = async (request: NextRequest) => {
@@ -17,9 +11,9 @@ export const GET = async (request: NextRequest) => {
 
     const data = goods;
 
-    const mappedData = mappingResponse(data);
+    const mappedData = mappingGoodsResponse(data);
 
-    const orderedData = orderingData(mappedData, order as Order);
+    const orderedData = orderingGoodsData(mappedData, order as Order);
 
     const pageSize = 20;
 
@@ -44,93 +38,4 @@ const parseSearchParams = (params: URLSearchParams): GetGoodsRequestParams => {
     order: (params.get("order") ?? "recommended") as Order,
     page: Number(params.get("page") ?? "1"),
   };
-};
-
-export const mappingResponse = (data: RawGoodDTO[]): GetGoodsResponse => {
-  const goods = data?.map((item) => {
-    return {
-      ...item,
-      price: {
-        ...item.price,
-        originalPriceDisplayText: formatNumberWithCommas(
-          Number(item.price.originalPrice) ?? "0"
-        ),
-        sellingPriceDisplayText: formatNumberWithCommas(
-          Number(item.price.sellingPrice) ?? "0"
-        ),
-      },
-      reviewStatistic: {
-        ...item.reviewStatistic,
-        reviewCountDisplayText:
-          formatNumberWithCommas(item.reviewStatistic.reviewCount) ?? "0",
-        reviewAverageDisplayText:
-          numberRounded(item.reviewStatistic.reviewAverage) ?? 0,
-      },
-    };
-  });
-
-  return {
-    goods,
-    totalResults: goods.length,
-  };
-};
-
-export const orderingData = (
-  data: GetGoodsResponse,
-  order: Order
-): GetGoodsResponse => {
-  if (order === "priceAsc") {
-    // 가격 오름차순 정렬
-    const sortedData = {
-      goods: data.goods.sort(
-        (a, b) => Number(a.price.sellingPrice) - Number(b.price.sellingPrice)
-      ),
-      totalResults: data.totalResults,
-    };
-
-    return sortedData;
-  } else if (order === "priceDesc") {
-    // 가격 내림차순 정렬
-    const sortedData = {
-      goods: data.goods.sort(
-        (a, b) => Number(b.price.sellingPrice) - Number(a.price.sellingPrice)
-      ),
-      totalResults: data.totalResults,
-    };
-
-    return sortedData;
-  } else if (order === "discountRate") {
-    // 할인율 내림차순 정렬
-    const sortedData = {
-      goods: data.goods.sort(
-        (a, b) => Number(b.price.discountRate) - Number(a.price.discountRate)
-      ),
-      totalResults: data.totalResults,
-    };
-
-    return sortedData;
-  } else if (order === "reviewAverage") {
-    // 평점 내림차순 정렬
-    const sortedData = {
-      goods: data.goods.sort(
-        (a, b) =>
-          b.reviewStatistic.reviewAverage - a.reviewStatistic.reviewAverage
-      ),
-      totalResults: data.totalResults,
-    };
-
-    return sortedData;
-  } else if (order === "reviewCount") {
-    // 리뷰 수 내림차수 정렬
-    const sortedData = {
-      goods: data.goods.sort(
-        (a, b) => b.reviewStatistic.reviewCount - a.reviewStatistic.reviewCount
-      ),
-      totalResults: data.totalResults,
-    };
-
-    return sortedData;
-  } else {
-    return data;
-  }
 };
